% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_api.R
\name{recall_api}
\alias{recall_api}
\title{recall_api}
\usage{
recall_api(search = NULL, lang = NULL, cat = NULL, lim = NULL)
}
\arguments{
\item{search}{string variable to search the database for}

\item{lang}{string variable as 'en' = English, 'fr' = French,
for response language}

\item{cat}{integer variable as 1 = FOOD, 2 = VEHICLES,
3 = HEALTH PRODUCTS, 4 = CONSUMER PRODUCTS}

\item{lim}{integer variable to limit results to n}
}
\description{
A function to return recall data from the Canadian Recalls and Safety Alerts API
}
\details{
The response is a dataframe
}
\examples{
recall_api('banana', 'en', 1, 10)
## will return 10 recall results for 'banana' in english

}
