% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_api_date.R
\name{recall_api_date}
\alias{recall_api_date}
\title{recall_api_date}
\usage{
recall_api_date(search = NULL, lang = NULL, cat = NULL,
  datestart = NULL, dateend = NULL)
}
\arguments{
\item{search}{string variable to search the database for}

\item{lang}{string variable as 'en' = English, 'fr' = French,
for response language}

\item{cat}{integer variable as 1 = FOOD, 2 = VEHICLES,
3 = HEALTH PRODUCTS, 4 = CONSUMER PRODUCTS}

\item{datestart}{string variable to set earliest date in range}

\item{dateend}{string ariable to set latest date in range}
}
\description{
A function to return recall data from the Canadian Recalls and Safety Alerts API, parsed by date range
dev
The result is a dataframe
}
\examples{
recall_api_date('banana', 'en', 1, '1/1/2019', '12/31/2019')
## will return all recall results for 'banana' in 2019 in english

}
