% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_api_recent.R
\name{recall_api_recent}
\alias{recall_api_recent}
\title{recall_api_recent}
\usage{
recall_api_recent(lang = NULL)
}
\arguments{
\item{lang}{string variable as 'en' = English, 'fr' = French,
for response language}
}
\description{
A function to return recall data from the Canadian Recalls and Safety Alerts API, parsed by date range
dev
The result is a dataframe
}
\examples{
recall_api_recent('en')
## will return 15 most recent recalls for all categories

}
